+++ CHANGELOG.md
@@ -9,6 +9,7 @@ CHANGES:
 * Remove type=str from command line parameters as this is the default for python. ([@TGWolf][])
 * Upgrade Available is now displayed as True/False instead of 1/0. ([@TGWolf][])
 * README.md added to each script to give more detailed overview. ([@TGWolf][])
+* Bug found in check-for-solution-stack-upgrade.py due to unsafe assumption, change in the lookup logic. ([@TGWolf][])
 
 ## 1.0.1 (November 6, 2018)